Restore back them to old place, for better diff
